Implement blueprints for sensors
