+++ website/app.py
@@ -9,6 +9,7 @@ from flask_utils import check_for_login
 
 from user_blueprint import user_blueprint
 from cab_blueprint import cab_blueprint
+from sensor_blueprint import sensor_blueprint
 
 app = Flask(__name__)
 app.static_folder = 'static'
@@ -16,6 +17,7 @@ app.static_folder = 'static'
 app.register_blueprint(client, url_prefix='')
 app.register_blueprint(user_blueprint, url_prefix='')
 app.register_blueprint(cab_blueprint, url_prefix='')
+app.register_blueprint(sensor_blueprint, url_prefix='')
 init_mqtt(app)
 
 
@@ -53,122 +55,6 @@ def home_page():
         return login_check
     print("Current user:", current_user, "Admin mode:", admin_mode)
     return render_template("home_page.html", current_user = current_user,admin_mode=admin_mode, turnouts = turnouts)
-
-# Handle CRUD for sensors
-
-@app.route('/try-register-sensor', methods=['POST', 'GET'])
-def try_register_sensor():
-    login_check = check_for_login()
-    if login_check != None:
-        return login_check
-    if request.method == 'POST':
-        sensor_id = request.form['id']
-        name = request.form['name']
-        value = request.form['value']
-        data = jsonutil.import_json(app.root_path + '/database/sensors.json')
-
-        if name.strip('') == "" or value.strip('') == "":
-            print("Sensor inválido!")
-            return redirect(app.url_for('register_sensor'))
-
-        for sensor in data:
-            if sensor['id'] == sensor_id:
-                print("Sensor já cadastrado!")
-                return register_sensor(error=True)
-        
-        data.append({'id': sensor_id, 'sensor': name, 'value': value})
-        jsonutil.export_json(app.root_path + '/database/sensors.json', data)
-        return redirect(app.url_for('list_sensor')) # Redirect to list of sensors later
-    
-
-@app.route('/try-edit-sensor', methods=['POST', 'GET'])
-def try_edit_sensor():
-    login_check = check_for_login()
-    if login_check != None:
-        return login_check
-    if request.method == 'POST':
-        #print("\n"*100, request.form)
-        edit_id = int(request.form['edit_id']   )
-        id = request.form['id']
-        name = request.form['name']
-        value = request.form['value']
-
-        data = jsonutil.import_json(app.root_path + '/database/sensors.json')
-
-        if id.strip('') == "" or name.strip('') == "" or value.strip('') == "" or not id.isnumeric():
-            print("Sensor inválido!")
-            return register_sensor(error=True, data=[id, name, value], edit_id=edit_id)
-
-        id = int(id)
-
-        index = 0
-        for sensor in data:
-            if index != edit_id and int(sensor['id']) == id:
-                print("Sensor já cadastrado!") 
-                return register_sensor(error=True, data=[id, name, value], edit_id=edit_id)
-            index += 1
-            
-        data[edit_id] = {'id': id, 'sensor': name, 'value': value}
-        jsonutil.export_json(app.root_path + '/database/sensors.json', data)
-        return redirect(app.url_for('list_sensor')) # Redirect to list of locomotives later
-    else:
-        print("Método inválido:", request.method)
-        return redirect(app.url_for('register_sensor')) # ERRO!
-
-@app.route('/edit-sensor', methods=['POST', 'GET'])
-def edit_sensor(error = False):
-    login_check = check_for_login()
-    if login_check != None:
-        return login_check
-    if request.method == 'POST':
-        sensor_id = request.form['sensor_id']
-        data = jsonutil.import_json(app.root_path + '/database/sensors.json')[int(sensor_id)]
-        return register_sensor(error, [data['id'], data['sensor'], data['value']], int(sensor_id))
-    else:
-        print("Método inválido:", request.method)
-        return redirect(app.url_for('register_sensor'))
-
-@app.route('/register-sensor')
-def register_sensor(error = False, data = ['', '', ''], edit_id = -1):
-    login_check = check_for_login()
-    if login_check != None:
-        return login_check
-    return render_template("register_sensor.html", error = error, data=data, edit_id = edit_id)
-
-@app.route('/remove-sensor')
-def remove_sensor():
-    login_check = check_for_login()
-    if login_check != None:
-        return login_check
-    return render_template("remove_sensor.html")
-
-@app.route('/list-sensor')
-def list_sensor():
-    login_check = check_for_login()
-    if login_check != None:
-        return login_check
-    data = jsonutil.import_json(app.root_path + '/database/sensors.json')
-    sensors = {}
-    index = 0
-    for sensor in data:
-        #print('\n\n\n\n\n\n\n\n\n\n\n\n\n', sensor)
-        sensors.update({index: sensor['sensor']})
-        index += 1
-    return render_template("list_sensor.html", sensors = sensors)
-
-@app.route('/try-remove-sensor', methods=['POST', 'GET'])
-def try_remove_sensor():
-    login_check = check_for_login()
-    if login_check != None:
-        return login_check
-    if request.method == 'POST':
-        sensor_id = request.form['sensor_id']
-        data = jsonutil.import_json(app.root_path + '/database/sensors.json')
-        # print("\n\n\n\n\n\n\n\n\n\nSensor a ser removido:", sensor_id)
-        name = data.pop(int(sensor_id))
-        jsonutil.export_json(app.root_path + '/database/sensors.json', data)
-        print("O sensor", name, "foi removido com sucesso!")
-        return redirect(app.url_for('list_sensor'))
 
 # Handle CRUD for detours
 
@@ -226,7 +112,7 @@ def try_edit_detour():
 
         if id.strip('') == "" or actuator.strip('') == "" or value.strip('') == "" or not id.isnumeric() or not value.isnumeric():
             print("Sensor inválido!")
-            return register_sensor(error=True, data=[id, actuator, value], edit_id=edit_id)
+            return register_detour(error=True, data=[id, actuator, value], edit_id=edit_id)
 
         id = int(id)
 
@@ -234,7 +120,7 @@ def try_edit_detour():
         for sensor in data:
             if index != edit_id and int(sensor['id']) == id:
                 print("Desvio já cadastrado!") 
-                return register_sensor(error=True, data=[id, actuator, value], edit_id=edit_id)
+                return register_detour(error=True, data=[id, actuator, value], edit_id=edit_id)
             index += 1
             
         data[edit_id] = {'id': int(id), 'actuator': actuator, 'value': int(value)}
