+++ website/templates/home_page.html
@@ -41,7 +41,7 @@
               </li>
               <li>
                 <!-- <i class="ph ph-broadcast"></i> -->
-                <a href="{{ url_for('list_sensor') }}">Sensores</a>
+                <a href="{{ url_for('sensor_blueprint.list_sensor') }}">Sensores</a>
               </li>
             {% endif %}
             <li>
